Move biomass species merge code to BiomassStore
The biomass species merge code was already neatly separated into its own method;
move that method to `BiomassStore` and move the `mergeOtherSpecies` entrypoint
method to `ObservationService` since it needs to call both the biomass and
plant-monitoring versions of the species merge code depending on observation type.
