+++ src/main/kotlin/com/terraformation/backend/tracking/db/BiomassStore.kt
@@ -4,6 +4,7 @@ import com.terraformation.backend.customer.db.ParentStore
 import com.terraformation.backend.customer.model.requirePermissions
 import com.terraformation.backend.db.asNonNullable
 import com.terraformation.backend.db.attach
+import com.terraformation.backend.db.default_schema.SpeciesId
 import com.terraformation.backend.db.tracking.BiomassForestType
 import com.terraformation.backend.db.tracking.MonitoringPlotId
 import com.terraformation.backend.db.tracking.ObservationId
@@ -15,8 +16,11 @@ import com.terraformation.backend.db.tracking.tables.records.ObservationBiomassQ
 import com.terraformation.backend.db.tracking.tables.records.ObservationBiomassQuadratSpeciesRecord
 import com.terraformation.backend.db.tracking.tables.records.ObservationBiomassSpeciesRecord
 import com.terraformation.backend.db.tracking.tables.records.RecordedTreesRecord
+import com.terraformation.backend.db.tracking.tables.references.OBSERVATION_BIOMASS_QUADRAT_SPECIES
 import com.terraformation.backend.db.tracking.tables.references.OBSERVATION_BIOMASS_SPECIES
 import com.terraformation.backend.db.tracking.tables.references.OBSERVATION_PLOTS
+import com.terraformation.backend.db.tracking.tables.references.RECORDED_TREES
+import com.terraformation.backend.log.perClassLogger
 import com.terraformation.backend.tracking.event.BiomassDetailsCreatedEvent
 import com.terraformation.backend.tracking.event.BiomassQuadratCreatedEvent
 import com.terraformation.backend.tracking.event.BiomassSpeciesCreatedEvent
@@ -25,6 +29,7 @@ import com.terraformation.backend.tracking.model.BiomassSpeciesKey
 import com.terraformation.backend.tracking.model.NewBiomassDetailsModel
 import jakarta.inject.Named
 import org.jooq.DSLContext
+import org.jooq.impl.DSL
 import org.springframework.context.ApplicationEventPublisher
 
 @Named
@@ -33,6 +38,8 @@ class BiomassStore(
     private val eventPublisher: ApplicationEventPublisher,
     private val parentStore: ParentStore,
 ) {
+  private val log = perClassLogger()
+
   fun insertBiomassDetails(
       observationId: ObservationId,
       plotId: MonitoringPlotId,
@@ -274,5 +281,126 @@ class BiomassStore(
         )
       }
     }
+  }
+
+  @Deprecated("Call this on ObservationService instead.")
+  fun mergeOtherSpecies(
+      observationId: ObservationId,
+      otherSpeciesName: String,
+      speciesId: SpeciesId,
+  ) {
+
+    val otherBiomassSpeciesId =
+        with(OBSERVATION_BIOMASS_SPECIES) {
+          dslContext.fetchValue(
+              ID,
+              SCIENTIFIC_NAME.eq(otherSpeciesName).and(OBSERVATION_ID.eq(observationId)),
+          )
+        }
+
+    if (otherBiomassSpeciesId == null) {
+      log.warn(
+          "Biomass observation $observationId does not contain species name $otherSpeciesName; " +
+              "merge is a no-op"
+      )
+      return
+    }
+
+    val targetBiomassSpeciesId =
+        with(OBSERVATION_BIOMASS_SPECIES) {
+          dslContext.fetchValue(ID, SPECIES_ID.eq(speciesId).and(OBSERVATION_ID.eq(observationId)))
+        }
+
+    if (targetBiomassSpeciesId == null) {
+      // The target species wasn't present at all in the observation, so there's no need to merge
+      // anything: we can just update the biomass species to point to the target species ID instead
+      // of using a name.
+      with(OBSERVATION_BIOMASS_SPECIES) {
+        dslContext
+            .update(OBSERVATION_BIOMASS_SPECIES)
+            .set(SPECIES_ID, speciesId)
+            .setNull(SCIENTIFIC_NAME)
+            .where(ID.eq(otherBiomassSpeciesId))
+            .execute()
+      }
+
+      return
+    }
+
+    // Recorded trees are a simple replacement of the biomass species ID.
+    with(RECORDED_TREES) {
+      dslContext
+          .update(RECORDED_TREES)
+          .set(BIOMASS_SPECIES_ID, targetBiomassSpeciesId)
+          .where(BIOMASS_SPECIES_ID.eq(otherBiomassSpeciesId))
+          .execute()
+    }
+
+    // For quadrat species, we need to add the abundance percentages of the numbered species and the
+    // "Other" one if both exist in the quadrat. If only the "Other" one exists, then we can just
+    // switch its biomass species ID in place.
+    with(OBSERVATION_BIOMASS_QUADRAT_SPECIES) {
+      val quadratSpeciesTable2 = OBSERVATION_BIOMASS_QUADRAT_SPECIES.`as`("quadrat2")
+
+      dslContext
+          .update(OBSERVATION_BIOMASS_QUADRAT_SPECIES)
+          .set(BIOMASS_SPECIES_ID, targetBiomassSpeciesId)
+          .where(BIOMASS_SPECIES_ID.eq(otherBiomassSpeciesId))
+          .andNotExists(
+              DSL.selectOne()
+                  .from(quadratSpeciesTable2)
+                  .where(quadratSpeciesTable2.BIOMASS_SPECIES_ID.eq(targetBiomassSpeciesId))
+                  .and(quadratSpeciesTable2.POSITION_ID.eq(POSITION_ID))
+          )
+          .execute()
+
+      dslContext
+          .update(OBSERVATION_BIOMASS_QUADRAT_SPECIES)
+          .set(
+              ABUNDANCE_PERCENT,
+              ABUNDANCE_PERCENT.plus(
+                  DSL.field(
+                      DSL.select(quadratSpeciesTable2.ABUNDANCE_PERCENT)
+                          .from(quadratSpeciesTable2)
+                          .where(quadratSpeciesTable2.BIOMASS_SPECIES_ID.eq(otherBiomassSpeciesId))
+                          .and(quadratSpeciesTable2.POSITION_ID.eq(POSITION_ID))
+                  )
+              ),
+          )
+          .where(BIOMASS_SPECIES_ID.eq(targetBiomassSpeciesId))
+          .andExists(
+              DSL.selectOne()
+                  .from(quadratSpeciesTable2)
+                  .where(quadratSpeciesTable2.BIOMASS_SPECIES_ID.eq(otherBiomassSpeciesId))
+                  .and(quadratSpeciesTable2.POSITION_ID.eq(POSITION_ID))
+          )
+          .execute()
+    }
+
+    // Invasive and threatened should be true if they're true on either version of the species.
+    with(OBSERVATION_BIOMASS_SPECIES) {
+      dslContext
+          .update(OBSERVATION_BIOMASS_SPECIES)
+          .set(
+              IS_INVASIVE,
+              DSL.select(DSL.boolOr(IS_INVASIVE))
+                  .from(OBSERVATION_BIOMASS_SPECIES)
+                  .where(ID.`in`(otherBiomassSpeciesId, targetBiomassSpeciesId)),
+          )
+          .set(
+              IS_THREATENED,
+              DSL.select(DSL.boolOr(IS_THREATENED))
+                  .from(OBSERVATION_BIOMASS_SPECIES)
+                  .where(ID.`in`(otherBiomassSpeciesId, targetBiomassSpeciesId)),
+          )
+          .where(ID.eq(targetBiomassSpeciesId))
+          .execute()
+    }
+
+    // ON DELETE CASCADE will remove the data for the "Other" species from all the biomass tables.
+    dslContext
+        .deleteFrom(OBSERVATION_BIOMASS_SPECIES)
+        .where(OBSERVATION_BIOMASS_SPECIES.ID.eq(otherBiomassSpeciesId))
+        .execute()
   }
 }
